refactor: create/styled -> common/styled
