+++ src/components/surveys/common/styled.js
