+++ src/components/surveys/create-card/DetailSurveyItem.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useRecoilState } from "recoil";
 import { detailMCQuestionState } from "../../../atoms";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { DetailContainer, Input } from "../create/styled";
+import { DetailContainer, Input } from "../common/styled";
 
 export default function DetailSurveyItem({ item }) {
   const [detailQuestionList, setDetailQuestionList] = useRecoilState(
