+++ src/components/surveys/create-card/PreviewSection.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Preview, TitleText, SummaryText } from "../create/styled";
+import { Preview, TitleText, SummaryText } from "../common/styled";
 import { useRecoilValue } from "recoil";
 import { surveyTitle, surveySummary, surveyListState } from "../../../atoms";
 import SurveyPreviewItem from "./SurveyPreviewItem";
