+++ src/components/surveys/create/NavBar.jsx
@@ -35,7 +35,7 @@ import {
 import { userState } from "../../../authentication/userState";
 import { CustomSwitch } from "./CustomizedSwitches";
 import SelectCategory from "./SelectCategory";
-import { SaveBtn, SNavBar } from "./styled";
+import { SaveBtn, SNavBar } from "../common/styled";
 import SurveyImg from "./SurveyImg";
 import { getAccessToken, getRefreshToken } from "../../../authentication/auth";
